+++ src/CodeGenerator/src/Generator/InputGenerator.php
@@ -595,11 +595,8 @@ PHP
                     throw new \InvalidArgumentException('Constant "\DateTimeInterface::' . $format . '" does not exists.');
                 }
 
-                if ('RFC822' === $format) {
-                    $variable .= '->setTimezone(new \\DateTimeZone("GMT"))';
-
-                    // Use RFC1123 instead of deprecated RFC7231 constant (PHP 8.5+)
-                    return $variable . '->format(\DateTimeInterface::RFC1123)';
+                if ('RFC822' === $format || 'RFC7231' === $format) {
+                    return '->setTimezone(new \\DateTimeZone("GMT"))->format(\'D, d M Y H:i:s \G\M\T\')';
                 }
 
                 return $variable . '->format(\DateTimeInterface::' . $format . ')';